refactor: server config api use
