+++ server/api/configuration.get.ts
