+++ utils/api/useServerConfiguration.ts
@@ -1,0 +1,14 @@
+export async function useServerConfiguration () {
+  const { data: config } = await useFetch<{
+    roles: Array<string>;
+    seniorityLevels: Array<string>;
+    departments: Array<string>;
+    currency: string;
+  }>('/api/configuration', {
+    lazy: true,
+  });
+
+  return {
+    config,
+  };
+}
