+++ utils/useToastNotifications.ts
@@ -1,0 +1,24 @@
+export function useToastNotifications () {
+  const toast = useToast();
+
+  function showErrorToast (error: string, description: string) {
+    toast.add({
+      title: error,
+      description,
+      color: 'error',
+    });
+  }
+
+  function showSuccessToast (title: string, description: string) {
+    toast.add({
+      title,
+      description,
+      color: 'success',
+    });
+  }
+
+  return {
+    showErrorToast,
+    showSuccessToast,
+  };
+}
